requirements.txt
# fastapi libs
uvicorn
fastapi~=0.111.0
pydantic~=2.7.1
starlette~=0.37.2
pydantic-settings~=2.2.1
gunicorn~=22.0.0
fastapi-event~=0.1.3

# Common libs
python-dotenv~=1.0.1
loguru~=0.7.0
requests~=2.31.0
dnspython~=2.6.1
supervisor~=4.2.5
jsonschema~=4.22.0
sqlalchemy
Werkzeug~=3.0.3
python-multipart~=0.0.9
pillow~=10.3.0

pytz~=2024.1
arrow~=1.3.0
APScheduler~=3.10.4









